Mappers: implemented Mapper001 (MMC1)
- only support 256kb mode for PRG size and not 512kb

+++ cartridge/src/mapper.rs
@@ -2,6 +2,6 @@ use common::Device;
 
 pub trait Mapper {
     fn init(&mut self, pgr_count: u8, chr_count: u8);
-    fn map_read(&self, address: u16, device: Device) -> u16;
+    fn map_read(&self, address: u16, device: Device) -> usize;
     fn map_write(&mut self, address: u16, data: u8, device: Device);
 }
